Add docs for io_util
