+++ Cargo.toml
@@ -65,3 +65,4 @@ rand = "0.8"
 sha2 = "0.10"
 size = "0.1"
 uuid = { version = "0.8", features = ["serde", "v4"] }
+serde_json = "1.0.79"
